Implemented move assignement and constructor for async.
